+++ app/Models/NavigationItem.php
@@ -20,6 +20,8 @@ class NavigationItem extends Model
 
     public const MENU_LEGAL = 'legal';
 
+    public const MENU_DOCS = 'docs';
+
     public const LINK_PAGE = 'page';
 
     public const LINK_CUSTOM_URL = 'custom_url';
@@ -30,6 +32,20 @@ class NavigationItem extends Model
 
     public const VISIBILITY_HIDDEN = 'hidden';
 
+    public const SIDEBAR_ICONS = [
+        'home',
+        'rocket',
+        'layers',
+        'palette',
+        'layout',
+        'box',
+        'star',
+        'grid',
+        'wrench',
+        'code',
+        'terminal',
+    ];
+
     protected $fillable = [
         'site_id',
         'menu_key',
@@ -39,6 +55,7 @@ class NavigationItem extends Model
         'link_type',
         'url',
         'target',
+        'icon',
         'position',
         'visibility',
         'is_system',
@@ -83,7 +100,7 @@ class NavigationItem extends Model
 
     public static function menuKeys(): array
     {
-        return [self::MENU_PRIMARY, self::MENU_FOOTER, self::MENU_MOBILE, self::MENU_LEGAL];
+        return [self::MENU_PRIMARY, self::MENU_FOOTER, self::MENU_MOBILE, self::MENU_LEGAL, self::MENU_DOCS];
     }
 
     public static function menuOptions(): array
@@ -93,6 +110,7 @@ class NavigationItem extends Model
             self::MENU_FOOTER => 'Footer',
             self::MENU_MOBILE => 'Mobile',
             self::MENU_LEGAL => 'Legal',
+            self::MENU_DOCS => 'Docs',
         ];
     }
 
@@ -109,6 +127,11 @@ class NavigationItem extends Model
     public static function locations(): array
     {
         return self::menuKeys();
+    }
+
+    public static function sidebarIconKeys(): array
+    {
+        return self::SIDEBAR_ICONS;
     }
 
     public function page(): BelongsTo
@@ -231,5 +254,12 @@ class NavigationItem extends Model
     public function isVisible(): bool
     {
         return $this->visibility !== self::VISIBILITY_HIDDEN;
+    }
+
+    public function sidebarIcon(): ?string
+    {
+        return in_array($this->icon, self::sidebarIconKeys(), true)
+            ? $this->icon
+            : null;
     }
 }